requirements.txt
cachetools==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
google-auth==1.0.2
google-cloud-bigquery==0.26.0
google-cloud-core==0.26.0
google-cloud-storage==1.3.1
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
idna==2.5
monotonic==1.3
protobuf==3.3.0
py==1.4.34
pyasn1==0.3.2
pyasn1-modules==0.0.11
pytest==3.2.0
requests==2.18.3
rsa==3.4.2
six==1.10.0
tenacity==4.4.0
urllib3==1.22
